shared/idmap: Split ByHostid
